rename: GlobalTracer to GlobalTracerDelegate
