+++ packages/opentelemetry-core/src/index.ts
@@ -16,7 +16,7 @@
 
 export * from './context/propagation/HttpTraceContext';
 export * from './resources/Resource';
-export * from './trace/GlobalTracer';
+export * from './trace/GlobalTracerDelegate';
 export * from './trace/globaltracer-utils';
 export * from './trace/NoopSpan';
 export * from './trace/NoopTracer';
